requirements.txt
#sentence-transformers==5.0.0    #open source vector embedding model
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
-e .